requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.7.0
Flask==3.0.3
flask-cors==6.0.0
Flask-Limiter==3.8.0
Flask-WTF==1.2.2
frozenlist==1.4.1
future==1.0.0
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
multidict==6.0.5
packaging==25.0
pycparser==2.22
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.2.1
yarl==1.9.4
psutil==5.9.8
line-bot-sdk